All css-like settings. Augment variable axes with transforms

+++ source/blender/makesdna/DNA_userdef_types.h
@@ -82,7 +82,7 @@ typedef struct uiFontStyle {
   /** 1 value, typically white or black anyway. */
   float shadowcolor;
   char _pad2[4];
-  float character_weight;
+  int character_weight;
   float character_slant;
   float character_width;
   float character_spacing;
